20181021 add product type

+++ src/main/java/cn/meilituibian/api/domain/Product.java
@@ -11,7 +11,7 @@ public class Product implements Serializable{
     private String content;
     private BigDecimal price;
     private int quantity;
-    private Long type;
+    private int type;
     private String typeName;
     private int status;  //0:无效,1:有效
     private String productImgs;
@@ -97,7 +97,7 @@ public class Product implements Serializable{
         this.version = version;
     }
 
-    public void setType(Long type) {
+    public void setType(int type) {
         this.type = type;
     }
